Rudimentary communication with a VNC client.
The screen.py file has a cairo example in it that plots some things
on a surface, and then ignores it. And it has a VNC service that can
listen for connections and does some rudimentary messaging so that
clients can connect to it.

The two things are not currently attached, and the service has a bunch
of fixed values in it. But it's moving in the right direction.