Refactor all the client messages into separate handlers.
Separate handler functions mean that it's easier to see the
implementation of each message and we don't have a large set of
indents to contend with. It also means that their behaviour has
to be made relatively generalised to the main class, which isn't
ideal because it splits the implementation of features between
the message handler and the class. However, as the alternative
would be to have the implementation spread across the class, it's
not really that much of a loss.