Add support for pixel conversion formats, and framebuffer update.
We now support the pixel conversion formats that are supplied as
SetPixelFormat - at least, the ones that are true colour. There is
a generic conversion format that tries to return the right format
as requested, although it's slower than the pass-through format
they get if they request the natural format.