Refactor the client messages and regions into separate modules.
The client messages are a set of registrations that can be dispatched
during to functions. They're now in a separate module, with the
intention of not cluttering the main module that handles the server
and its main functionality. This does spread some of the implementation
about a bit, but I think it is easier to have this registration interface
which you can then hook things into when needed.