Clean up the server implementation.
The interface to start the server is now through a separate class
which controls the server object itself. This means should allow us
to start and stop the server arbitrarily, and we should be able
to introduce controls on the clients, etc. There's a bunch of
parameters we should be passing through the system so that we can
make the operations safe - locking and the like is still completely
lacking, as is any way to change the display size.