Initial support for security through authentication.
The authentication that is provided is only for the 'None' at the
moment. This change refactors the authentication code into an object
which can handle the negotiation. In the future it may also be able
to handle encryption of the passed data as well, but right now we
only support the 'no authentication' and don't touch any encryption
at all.

Extending this with VNCAuthentication will follow and should be
pretty simple.