Fix for non-32bit colour data.
When the client requested non-32bit data, eg a 16 bit rendering,
we would produce images that used the low bits of the colour data,
rather than the high bits. This was due to incorrectly shifting
and ANDing the data from our internal format.

We now shift the data around and AND the data correctly so that the
depth is correct for what they want to see.

Strictly there may be more updates to the data than the client
expects if they are reducing the quality, because the rows that
we send are only compared at the source. So if the colours are
changing very slightly there may be an update sent which produces
the same image as was there before. Short of converting the image
to what the client is actually seeing before processing the
differences, this is unavoidable.