Add support for VNC authentication.
VNC authentication is now supported using the DES library. This
needs to be installed separately (it's not duplicated here).
There are still problems with the Apple screen sharing - it doesn't
allow you to connect without password and just ignores the fact that
Security_None is being returned. However, when you give a password
it does connect.

Apart from the authentication the Apple screen sharing client also
doesn't solicit updates; it expects them to be sent unsolicited.