Refactor the initialisation phases into smaller functions.
The phases were inline in the handle function, which makes them a
more difficult to reason about and harder to extend. This change
moves each of these phases into separate methods which are then
invoked in order within the handle function.

This might make it easier for someone to override in a subclass
in the future, but it's probably not quite the right level to do
that just yet.