Add support for surface changing.
The underlying surface can now change. The surface that is in use
can change to be a different size, and if the client is aware of
the DesktopSize encoding, we'll notify it. This means a bit more
processing of messages in the system, but it seems to be working
correctly - the regions are updated at the point at which the surface
changes and we don't end up the clients becoming stuck.

There is a problem with the RealVNC client, which seems to disconnect
sometimes when the framebuffer changes in size. It's not clear why;
the other VNC clients that have been tested do not exhibit this
problem.

There's some tentative support for changing the desktop name as well,
but this is not yet complete.

The examples have been moved around, with separate examples for
the simple case and the case where the surface changes.