Only report authentication present when we can authenticate.
The 'None' authentication should work with all clients. That it does
not work with the macOS client should not be a reason to force
everyone to use a password - you might not want to. So I've
reverted the change which allowed the VNC authentication all the
time. macOS screen sharing is (I believe) the broken one here,
and whilst unauthenticated access might not be great for security,
it is the user's choice.