Add thread safety to the surface data.
The surface lock and the surface data lock are now implemented.

The surface lock makes access to the surface safe; the animator
can update the frames protected by a lock, and the VNC clients
can obtain it as a full frame.

The surface data lock ensures that only one of the VNC clients
is caching the data at a time - otherwise multiple clients
might be trying to pull the data from the surface, and starving
the animator (spoiling the purpose of our cache).

Example has been added which shows the use of the lock to ensure
that we have safe access to the surface.