Introduce throttling to framebuffer updates.
The framebuffer updates are now throttled so that the VNC client
isn't controlling how much CPU usage we have - we delay our responses
until the period for frames occurs.

In particular, more VNC clients will send a frame request as soon as
they receive the prior one. This results in the server being hit as
quickly as it can service requests. It hits 100% very easily, with
just one client connected. With the throttling present, however,
with the example code on a 2013 MacBook and 4 clients connected,
python only hits 4% CPU usage, and the clients are mostly smooth.