Add support for changing the desktop name.
The desktop name can now be changed whilst the server is running.
Changes to the name are only supported by some clients - the RealVNC
viewer does not appear to support the name change, for example.

This also fixes the name of the desktop being encoded in Latin-1;
apparently it's encouraged to encode it in UTF-8.