Improved handling of server shutdown.
When the server shut down it could leave clients blocked because
the event queue was full. This problem can be mitigated by clearing
the queue before we return, but that only prevents the clients
that are connected up to the length of the event queue from being
blocked. It's such an unlikely circumstance, that I'm not going
to worry too much about it.

Also the wrong function was being called for shutting down the
server, so it would fail to shutdown anyhow.