Limit clients; improve documentation of the methods.
The maximum number of clients can now be set, and any above that
will be dropped. This allows us to limit the resources which are
used for a given server.

The methods in the main class have been updated so that they now
have some better documentation to explain their usage - in particular
which threads the methods can be used on.